refactor : 로또 관련 공통 검증 로직을 분리
