+++ src/main/java/model/common/LottoConstants.java
@@ -1,4 +1,4 @@
-package model;
+package model.common;
 
 public class LottoConstants {
     public static final int MIN_NUMBER = 1;
