+++ src/main/java/model/purchase/Lotto.java
@@ -1,10 +1,9 @@
 package model.purchase;
 
-import exception.ExceptionMessage;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
-import model.LottoConstants;
+import model.common.LottoValidator;
 import model.draw.BonusNumber;
 import model.draw.WinningNumber;
 
@@ -12,11 +11,7 @@ public class Lotto {
     private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
-        validateNumberCount(numbers);
-        validateNumberDuplication(numbers);
-        for (int number : numbers) {
-            validateNumberRange(number);
-        }
+        validate(numbers);
         this.numbers = numbers;
     }
 
@@ -38,21 +33,11 @@ public class Lotto {
         return Collections.unmodifiableList(numbers);
     }
 
-    private void validateNumberCount(List<Integer> numbers) {
-        if (numbers.size() != LottoConstants.NUMBER_COUNT) {
-            throw new IllegalStateException(ExceptionMessage.OUT_OF_LOTTO_SIZE);
-        }
-    }
-
-    private void validateNumberDuplication(List<Integer> numbers) {
-        if (new HashSet<>(numbers).size() != numbers.size()) {
-            throw new IllegalArgumentException(ExceptionMessage.DUPLICATED_LOTTO_NUMBER);
-        }
-    }
-
-    private void validateNumberRange(int number) {
-        if (number < LottoConstants.MIN_NUMBER || number > LottoConstants.MAX_NUMBER) {
-            throw new IllegalStateException(ExceptionMessage.OUT_OF_LOTTO_NUMBER_RANGE);
+    private void validate(List<Integer> numbers) {
+        LottoValidator.validateLottoNumberCount(numbers);
+        LottoValidator.validateLottoNumberDuplication(numbers);
+        for (int number : numbers) {
+            LottoValidator.validateLottoNumberRange(number);
         }
     }
 }
