+++ src/test/java/model/draw/WinningStatusTest.java
@@ -12,7 +12,8 @@ class WinningStatusTest {
 
     @DisplayName("로또 번호와 당첨 번호의 일치 개수와 보너스 번호와의 일치 여부를 통해 당첨 결과를 구한다")
     @ParameterizedTest
-    @CsvSource(value = {"6:false:FIRST", "5:true:SECOND", "5:false:THIRD", "4:false:FOURTH", "3:false:FIFTH", "2:false:NONE", "1:false:NONE", "0:false:NONE"}, delimiter = ':'
+    @CsvSource(value = {"6:false:FIRST", "5:true:SECOND", "5:false:THIRD", "4:false:FOURTH", "3:false:FIFTH",
+            "2:false:NONE", "1:false:NONE", "0:false:NONE"}, delimiter = ':'
     )
     void findWinningStatus(int matchingCount, boolean matchesBonusNumber, WinningStatus expectedStatus) {
         assertThat(WinningStatus.decideBy(matchingCount, matchesBonusNumber)).isEqualTo(expectedStatus);
@@ -30,8 +31,8 @@ class WinningStatusTest {
     @Test
     void sortStatusByPrice() {
         List<WinningStatus> sortedWinningStatus = WinningStatus.getAllSortedByPrice();
-        for(int i = 0; i < sortedWinningStatus.size() - 1; i++) {
-            assertThat(sortedWinningStatus.get(i).getPrice()).isLessThan(sortedWinningStatus.get(i+1).getPrice());
+        for (int i = 0; i < sortedWinningStatus.size() - 1; i++) {
+            assertThat(sortedWinningStatus.get(i).getPrice()).isLessThan(sortedWinningStatus.get(i + 1).getPrice());
         }
     }
 }
